Building fix
Former-commit-id: f5f3d2b6bcba23144f9f0a74f259139bac13b859

+++ Sources/Internal/Network/SimpleNetworking/Private/SimpleTcpServer.cpp
@@ -86,7 +86,7 @@ bool SimpleTcpServer::Bind()
 {
     const sockaddr* addr = reinterpret_cast<const sockaddr*>(socketEndPoint.CastToSockaddrIn());
 
-    int bindRes = ::bind(socketId, addr, socketEndPoint.Size());
+    int bindRes = ::bind(socketId, addr, static_cast<int>(socketEndPoint.Size()));
     if (!CheckSocketResult(bindRes))
     {
         Close();
